ignore intra-emphasis unless followed by '<'
As reported here[1], the setting `no_intra_emphasis` has no effect when
the second underscore is followed by a character that is not in the set
[a-zA-Z]. This has the side effect that underscores are interpreted as
marking emphasis in strings like this:

    http://en.wikipedia.org/wiki/Dave_Allen_(comedian)

Any punctuation character following an underscore is considered to end
the potentially emphasised string, hence "Allen" in the above example
would be rendered as `<em>Allen</em>`. This commit changes this
condition, such that only a character in the "space" class or an opening
angular bracket (`<`) after an underscore results in emphasised text
when `no_intra_emphasis` is set.

[1]: https://github.com/tanoku/redcarpet/issues/142