Make ordered lists preceded by paragraph parsed
Previously, enabling the `:lax_spacing` option, if a paragraph was
followed by an ordered list it was previously unparsed and was part
of the paragraph but this is no more the case.

The fix is pretty straightforward, since the `parse_paragraph` method
was checking through the `isalnum` function whether the first char on
the next line was alpha-numeric but ordered list start with numerics,
we just need to substitute the function with isalpha.

Fixes #311