Improve Travis' feedback
Add 2.1.0 to the list of tested rubies but also keep ruby-head to be
sure to be compatible with the current trunk since Ruby moved to
semantic versioning, the release process is a bit faster.

Also simply test on the latest Rubinius version and allow failures on
it until we can find an appropriate fix for the code introduced in #333.

We should also rely on the travis_retry command when running bundle
install since the rake-compiler gem set up often generate a Bundler
error.