Change MKDEXT_NO_INTRA_EMPHASIS to not start after alphanumerics
No behavior should change if the NO_INTRA_EMPHASIS flag is not set.
With it
set, the following behavior changes should be seen:  (The parentheses
in the
following examples can be replaced by any non-alphanumeric character.)

- "(_italics_)" is italic
- "this(_is_)weird" italicizes the "is"

We were already checking for non-alphanumerics  on the closing boundary;
that is, "_this_()" would have rendered as italicized even though the
closing
underscore wasn't followed by a space. This commit just brings the same
behavior to the opening underscore.