Introduce _isalnum checking ASCII alphanumerics
`_isalnum` checks if the given character is an alphanumeric character
in the ASCII range. (The code is borrowed from github-markdow.)

This commit enables emphasis inside of sentences in multi-byte languages
when the the NO_INTRA_EMPHASIS flag is set.

For example, given the following sentence (in Chinese):

    中文內無空白的**粗體**也應加粗。

The two characters surrounded by asterisks should be bold.

This frees multi-byte language users from adding extra spaces into
sentences and is exactly how markdown on Github behaves.