Rely on #scan instead of #split
Let's rely on #scan in combination with #length instead of #split to
make the assertion more clear. Previously, we were asserting that the
array returned by #split have a length of n+1 where n is the actual
number of elements. With #scan, this is now a lot more easy to read.