Move the RedcarpetCompat class to its own file
Since the RedCloth API layer is now deprecated and will be removed in
4.0, let's move this class to its own file to make the main one more
clear.

We would have to require "redcarpet/compat" to be fully backward
compatible but the RedcarpetCompat class shouldn't be use anyway. It
is just a convenient way not to override the `Markdown` one if it
already exist.