Don't autolink text if it's part of a previous inline
Previously it was possible to get the renderer to truncate the
output at a point that would cause it to output unbalanced tags,
for ex.:

`__o__@bar.com` -> `<strong>o</st<a href="mailto:__o__@bar.com[…]`

Not great. This commit attempts to mitigate that by tracking the end
of the last inline and making sure rewinds don't run past it. This is
basically the approach taken by @vmg in rinku:

vmg/rinku@dc685c1