Strip out `style` tags from HTML blocks
The `rndr_raw_html` function is responsible for removing the `a`, `img`
and `style` tags when their respective :no_ option is enabled. However,
this function only acts inside paragraphs so `style` tags considered as
HTML blocks won't be removed from the output.

This is not a problem for links and images since they automatically get
included inside a paragraph since they're not considered as block
elements (c.f. html_blocks.h).

Thus, we need to duplicate the logic of the aforementioned function
inside the `rndr_raw_block` function.

Fixes #421.