Correctly strip out surrounding bad chars
Generating anchors, titles with chars that should be stripped out at the
beginning or at the end, used to have wrong anchors since bad chars were
replaced with a dash.

We just need to begin to add title's chars when we find a letter and
avoid replacing a bad char with a dash if it's the last one.

Also add a sharp (`#`) during the assertion to make sure that we match
starting from the very beginning of the anchor.