Avoid requiring the top level file in the 'compat' file
As of #401, the compatibility layer lives in its own file but since
Yard provides a Redcarpet support and relies on the same symbol to
specify the provider and the require string (https://git.io/vDLyr),
we can't ask them to "require 'redcarpet/compat'" so let's still
require the compatibility layer by default.

Fixes #589.