Correct an issue where the list parser would improperly handle indentation
By observing the data available at work->data, an interesting observation was noted that causes the tracking of indentation to be improperly handled. Specifically, the indentation on Item 3 appears to be handled incorrectly, which causes issues when scanning on the next iteration, from Item 2 downward.

To correct this, indentation is first handled as normal, checking up to four characters. We do this by initializing pre to our starting value, 4, then perform the indentation check using pre instead of the fixed value. Since pre is being updated constantly, this allows us to take the current indent and use it to properly crawl the list.

An interesting note is that the test for this issue will fail if you add a sixth-level item. This is due to max_nesting and how nesting is tracked on lists. On a linear list, a max_nesting of (1 + 2 + ... + n) is required, with n being the furthest nesting level. By default, max_nesting is 16, and a list 5 deep works, hitting a nesting of 15. However, with 6 levels, we need a max_nesting of 21 or greater.