Use Ruby's `ruby_xmalloc`
Ruby's `ruby_xmalloc` is defined here[1] and it calls `malloc` underneath.

The main advantage of using `ruby_xmalloc` over plain `malloc`, is that if the
memory allocation fails, an error will be raised instead of getting a `NULL`
pointer that is eventually going to be dereferenced, and thus, the process will
receive a `sigsegv`.

The implementation of this[2]

[1] https://github.com/ruby/ruby/blob/48ce4cb414be48be1a73d2530203469b3b98281c/gc.c#L7995
[2] https://github.com/ruby/ruby/blob/48ce4cb414be48be1a73d2530203469b3b98281c/gc.c#L7905-L7912