Allow 2-space indentation for multi-p li items
Other implementations of markdown allow for smaller indentations to
indicate the continuation of a list item, whereas the main redcarpet
repository enforces at least 4 spaces.

The particular motivation for this change was that we've switched our
curriculum repository to use mdformat, which always reformats list items
to match the level of indentation of the list indicator (2 for "* ", 3
for "1. " etc.). This will always be at least 2 spaces.

Note: This generally feels a bit hacky, in that the indentation is taken
relative to the LHS, rather than relative to the previous bullet point.
However, I'm not planning on fixing this unintuitive behaviour for now,
this small change should do what we need (and tests still pass).